# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ecossistema Whats API
  version: 0.1.0
servers:
  - url: https://ecossistemawhats.vercel.app
paths:
...
  /api/commands:
    post:
      summary: Ingest WhatsApp or Telegram command and create a microtask
      security:
        - bearerAuth: []
        - telegramSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, text]
              properties:
                channel:
                  type: string
                  enum: [whatsapp, telegram]
                from:
                  type: string
                  description: WhatsApp E.164 number
                chatId:
                  type: string
                  description: Telegram chat ID
                text:
                  type: string
                  examples: [/hoje, /estoque, /delegar conferir estoque amanha]
                contactName:
                  type: string
                payload:
                  type: object
      responses:
        "202":
          description: Command accepted and microtask created or simulated
        "401":
          description: Invalid bearer or Telegram secret
        "403":
          description: Sender not allowed
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
    webflowSignature:
      type: apiKey
      in: header
      name: x-webflow-signature
    telegramSecret:
      type: apiKey
      in: header
      name: x-telegram-bot-api-secret-token